Rework on history snapshots logic
